% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARISTAS-package.R
\docType{package}
\name{ARISTAS-package}
\alias{ARISTAS}
\alias{ARISTAS-package}
\title{ARISTAS: Análisis de los Resultados de los Datos ARISTAS del Instituto Nacional de Evaluación Educativa de Uruguay}
\description{
Este paquete analiza las bases de datos ARISTAS del Instituto Nacional de Evaluación Educativa de Uruguay. Para una edición específica de ARISTAS, el paquete realiza la estimación y pruebas de hipótesis de los principales indicadores educativos a nivel nacional y también según subgrupos definidos por el usuario. Para diferentes ediciones de ARISTAS, el paquete realiza comparaciones entre años. Además, el paquete ajusta modelos multinivel permitiendo tanto pesos al nivel de los estudiantes como al nivel de las escuelas.
}
\author{
\strong{Maintainer}: Hanwen Zhang \email{hanwenzhang1982@gmail.com}

}
\keyword{internal}
