% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARISTAS22.matematicas.R
\docType{data}
\name{ARISTAS22.matematicas}
\alias{ARISTAS22.matematicas}
\title{Datos ARISTAS lectura edición 2022}
\format{
\subsection{\code{ARISTAS22.matematicas}}{

Una base de datos con 11546 filas (estudiantes) y 415 columnas (variables)
\describe{
\item{AlumnoCodigoDes}{Identificador del estudiante}
\item{GrupoCodigoDes}{Identificador del grupo}
\item{CentroCodigoDes}{Identificador del centro educativo}
\item{MdeoInt}{Región ("Montevideo", "Interior")}
\item{regiones}{Región ("SUR", "OESTE", "NORTE", "ESTE", "CENTRO")}
\item{categoria_centro}{Tipo de Centro ("Secundaria", "pública", "Secundaria privada", "UTU (escuela técnica)")}
\item{ESCS_Centro}{Índice de contexto socioeconómico y cultural del centro}
\item{ESCS_Centro_cat}{Índice de contexto socioeconómico y cultural del centro en quintiles (Q1: "Muy desfavorable", Q2: "Desfavorable", Q3: "Medio", Q4: "Favorable", Q5: "Muy favorable")}
\item{ESCS_Centro_cat_comparable2018}{Índice de contexto socioeconómico y cultural del centro en quintiles comparables con Aristas 2018 (Q1: "Muy desfavorable", Q2: "Desfavorable", Q3: "Medio", Q4: "Favorable", Q5: "Muy favorable")}
\item{categoria_grupo}{Tipo de curso ("Secundaria pública", "Secundaria privada", "UTU-CBT (escuelas técnicas con ciclo básico tecnológico)", "UTU-FPB (escuelas técnicas con formación profesional básica)")}
\item{ESCS_Grupo}{Índice de contexto socioeconómico y cultural del grupo}
\item{ESCS_Grupo_cat}{Índice de contexto socioeconómico y cultural del grupo en quintiles (Q1: "Muy desfavorable", Q2: "Desfavorable", Q3: "Medio", Q4: "Favorable", Q5: "Muy favorable")}
\item{EDAD}{Edad}
\item{AlumnoGenero}{Sexo (F: "Femenino", M: "Masculino")}
\item{Concurre_cod}{Asistencia a clase reportada por el referente del grupo (1 "Asiste habitualmente", 2 "Asiste poco, todas las semanas falta algunos días",3 "Asiste muy poco, aproximadamente 1 vez cada 15 días , 4 "No asiste se cambió de centro", 5 "No asiste se cambió de grupo", 6 "No asiste, dejó de venir hace dos meses o más")}
\item{INSE}{Índice de nivel socioeconómico }
\item{ESCS_Alumno}{Estatus socioeconómico y cultural del estudiante}
\item{ESCS_Alumno_cat_comparable2018}{Estatus socioeconómico y cultural del estudiante en quintiles comparable con Aristas 2018 (Q1: "Muy desfavorable", Q2: "Desfavorable", Q3: "Medio", Q4: "Favorable", Q5: "Muy favorable")}
\item{ESCS_Alumno_cat}{Estatus socioeconómico y cultural del estudiante en quintiles (Q1: "Muy desfavorable", Q2: "Desfavorable", Q3: "Medio", Q4: "Favorable", Q5: "Muy favorable")}
\item{ESCS_Centro_comparable2018}{Índice de contexto socioeconómico y cultural del centro comparable con Aristas 2018}
\item{peso_CENTRO}{Ponderador de centro}
\item{peso_MGrup}{Ponderador de grupo}
\item{NEE}{Estudiantes identificados con necesidades educativas específicas cuyos resultados se excluyeron de las pruebas de Aristas (1 "estudiante con NEE excluido del puntaje de las pruebas de Aristas", 0 "estudiante no excluido del puntaje de las pruebas de Aristas")}
\item{ESCS_Alumno_comparable2018}{Estatus socioeconómico y cultural del estudiante comparable con 2018}
\item{theta_LEN_300_50}{Habilidad obtenida en la prueba de lectura}
\item{Niveles_LEN}{Nivel de desempeño en lectura (N1 "Nivel 1", N2 "Nivel 2", N3 "Nivel 3", N4 "Nivel 4", N5 "Nivel 5", N6 "Nivel 6")}
\item{theta_MAT_300_50}{Habilidad obtenida en la prueba de matemática}
\item{Niveles_MAT}{Nivel de desempeño en matemática (N1 "Nivel 1", N2 "Nivel 2", N3 "Nivel 3", N4 "Nivel 4", N5 "Nivel 5")}
\item{peso_MEst}{Ponderador del estudiante}
\item{EST_W_REP_1 : EST_W_REP_160}{Pesos repetidos mediante método BRR (Balanced Repeated Replication)}
\item{MAT1010, ..., MAT860}{Resultados a ítems de la prueba.}
\item{TotResp}{"Cantidad de ítems visualizados (sin contar los ítems eliminados en el proceso de calibración)}
\item{IND_Estudiante_ITEMS}{Respuesta al cuestionario (1: "El alumno hizo la prueba", vacía: "El alumno no hizo la prueba")}
}
}
}
\source{
\url{https://www.ineed.edu.uy/base-de-datos-aristas-media-2022/}
}
\usage{
ARISTAS22.matematicas
}
\description{
Aristas, la Evaluación Nacional de Logros Educativos, es realizada en Uruguay por el Instituto Nacional de Evaluación Educativa (INEEd).
Tiene como objetivo fundamental dar cuenta de los logros del sistema educativo en tercero y sexto de primaria y en tercero de educación media a nivel nacional en múltiples dimensiones. En el año 2022 se aplicó Aristas en tercero de media.
En esta base 'ARISTAS22.matematicas' se incluye la información de desempeño en matemáticas
}
\keyword{datasets}
