% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal.R
\name{longitudinal}
\alias{longitudinal}
\title{Esta función compara las diferentes ediciones de ARISTAS}
\usage{
longitudinal(
  datos,
  ediciones,
  y,
  x = NULL,
  digit = 3,
  plot = "density",
  descarga = FALSE,
  archivo = NULL,
  ...
)
}
\arguments{
\item{datos}{La base de datos que se quiere analizar, debe tener una columna llamada edicion indicando a cuál edición corresponde cada fila de la base.}

\item{ediciones}{Vector de caracteres o números indicando cuáles son las ediciones que desea comparar.}

\item{y}{El nombre de la variable de interés, ésta variable puede ser de tipo numérico o categórico.}

\item{x}{(Opcional) Nombre de la variable numérica o categórica que define los subgrupos para la comparación.}

\item{digit}{Especificar el número de dígitos decimales, por defecto son 3 dígitos decimales.}

\item{plot}{Puede elegir entre \code{density}, \code{violin}, \code{Lollipop}, \code{bar}. La opción \code{violin} incluirá también la gráfica de cajas.}

\item{descarga}{Indicar si desea descargar los resultados numéricos y la gráfica, debe especificar la ubicación. Al usar la opción \code{descarga = TRUE} se descargará un archivo excel con los resultados numéricos y un pdf con la gráfica.}

\item{archivo}{Indicar el nombre del archivo excel que se descarga. La gráfica, si hay, se descargará con el mismo nombre y extensión pdf en la misma ubicación que el archivo excel.}

\item{ancho}{El ancho en pulgadas de la gráfica.}

\item{alto}{El alto en pulgadas de la gráfica.}
}
\value{
\code{longitudinal}: devuelve una lista de objetos correspondientes a estimacions y resultados de hipótesis de comparación:
\itemize{
\item \code{estimaciones}: La estimación, error estándar, intervalo de confianza de 95\% e información de los NA para la media de la variable de interés para cada valor de la variable de corte \code{x} y para cada edición
\item \code{estimaciones.globales}: Las estimaciones, errores estándares, intervalos de confianza de 95\% e información de los NA para la media de la variable de interés para cada edición a nivel global (para todos los datos ingresados en el argumento \code{data}). Cuando no se ingresa variable de corte \code{x}, este resultado coincide con \code{estimaciones}, por lo que se omite en la salida.
\item \code{contraste.global}: Los resultados de prueba de hipótesis para la igualdad de medidas de la variable de interés en diferentes ediciones.
\item \code{contrastes.invididuales.entre.ediciones}: Los resultados de prueba de hipótesis para la igualdad de medidas de la variable de interés en cada valor de la variable de corte \code{x} en diferentes ediciones .
}

Adicionalmente, también la función arrojará la gráfica que usuario haya elegido.
}
\description{
Esta función compara las diferentes ediciones de ARISTAS
}
\details{
Cuando la variable de interés \code{y} es una variable numérica (puntaje en pruebas), y el usuario ingresa la varible de corte \code{x} con o sin valores en el argumento \code{limits},
la función realizará la prueba \eqn{F} para la igualdad de medias de la variable de interés \code{y} en cada categoría de \code{x} entre las ediciones, la prueba se realiza con la función \code{survey::svyglm} y \code{survey::regTermTest} con base en un diseño muestral de pesos réplica definido con \code{survey::svrepdesign},
y estos resultados se encuentran en el resultado \code{contraste.individual}. La salida de la función incluye la estimación, error estándar, intervalo de confianza de 95\% para la media de la variable \code{y} para cada categoría de \code{x} y a nivel global
También calcula el número de datos, número y porcentaje de NA para cada categoría de \code{x} y a nivel global para cada edición.
De la misma forma, la función llevará a cabo la prueba global \eqn{F} de igualdad de medias  entre las diferentes ediciones.

Cuando la variable de interés \code{y} es una variable numérica (puntaje en pruebas), y el usuario no ingresa la varible de corte \code{x},
la función solo calculará la estimación, error estándar, intervalo de confianza de 95\%, cantidad de datos, cantidad y porcentaje de NA para la media de la variable \code{y} para cada edición,
y también la prueba global \eqn{F} de igualdad de medias  entre las diferentes ediciones.

Cuando la variable de interés \code{y} es una variable categórica (nivel de desempeño), y el usuario ingresa la varible de corte \code{x} con o sin valores en el argumento \code{limits},
la función realizará la prueba chi-cuadrado para la independencia entre los diferentes valores de \code{y} y las ediciones para cada categoría de la variable \code{x} haciendo uso de la función \code{survey::svytable},
lo anterior con base en un diseño muestral de pesos réplica definido con \code{survey::svrepdesign},
y estos resultados se encuentran en \code{contrastes.invididuales.entre.ediciones}.
También calcula el número de datos, número y porcentaje de NA para cada categoría de \code{x} y a nivel global para cada edición.
De la misma forma, la función llevará a cabo la prueba global de independencia de chi-cuadrado entre los diferentes valores de \code{y} y las ediciones.

Cuando la variable de interés \code{y} es una variable categórica (nivel de desempeño), y el usuario no ingresa la varible de corte \code{x},
la función solo calculará la estimación, error estándar e intervalo de confianza de 95\%, cantidad de datos, cantidad y porcentaje de NA para la variable \code{y} en cada edición.

La opción de \code{plot} para una variable de interés \code{y} numérica admite \code{histogram}  (el cual hace uso de la función \code{survey::svyhist}), \code{density}, \code{violin} y \code{Lollipop}, mientras que
para una variable de interés \code{y} categórica admite la opción de \code{bar}.

Las estimaciones se obtuvieron utilizando la función \code{svymean} del paquete \code{survey} apoyado en un diseño muestral de pesos réplica creado con \code{svrepdesign}
}
\examples{
# ------------------------------------------------------------------------------------------
# Unificar categorías de las variables
# ------------------------------------------------------------------------------------------
# Convertir las variables numéricas a categóricas, la variable región está como entero!!!
ARISTAS18.matematicas$regiones <- as.character(ARISTAS18.matematicas$regiones)
# Unificar las categorías de ARISTAS 2018 según las categorías de 2022
ARISTAS18.matematicas2 <- ARISTAS18.matematicas |>
  dplyr::mutate(regiones = dplyr::case_when(
    regiones == 1 ~ "SUR",
    regiones == 2 ~ "OESTE",
    regiones == 3 ~ "NORTE",
    regiones == 4 ~ "ESTE",
    regiones == 5 ~ "CENTRO",
    TRUE ~ NA_character_  # Para asegurarse de que otros valores no definidos se traten como NA
    )
  )

# -------------------------------------------------------------------------------------------------
# Comparación de puntaje de matemáticas
# -------------------------------------------------------------------------------------------------
 # Unificar el nombre de la variable theta_MAT_300_50
ARISTAS18.matematicas2 <- ARISTAS18.matematicas2 |>
  dplyr::rename(theta_MAT_300_50 = theta_MAT_E300,
                AlumnoGenero = AlumnoGenero15)


 # Unir las bases de las dos ediciones a una sola base de datos
ARISTAS18.22.matematicas <- unionbases.longitudinal(ediciones = c(2018, 2022), bases = c("ARISTAS18.matematicas2", "ARISTAS22.matematicas"), y = "theta_MAT_300_50", x = "regiones")

# Análisis de la variable theta_MAT_E300 por región a nivel nacional
longitudinal(datos = ARISTAS18.22.matematicas, ediciones = c(2018, 2022), y = "theta_MAT_300_50", digit = 3, plot = "violin")

# Análisis de la variable theta_MAT_E300 por región y edición
longitudinal(datos = ARISTAS18.22.matematicas, ediciones = c(2018, 2022), y = "theta_MAT_300_50", x = "regiones", digit = 1, plot = "density")
longitudinal(datos = ARISTAS18.22.matematicas, ediciones = c(2018, 2022), y = "theta_MAT_300_50", x = "regiones", digit = 2, plot = "Lollipop")
longitudinal(datos = ARISTAS18.22.matematicas, ediciones = c(2018, 2022), y = "theta_MAT_300_50", x = "regiones", digit = 3, plot = "violin", descarga = TRUE, archivo = "theta_MAT_regiones.xlsx")
longitudinal(datos = ARISTAS18.22.matematicas, ediciones = c(2018, 2022), y = "theta_MAT_300_50", x = "regiones", digit = 4, plot = "density", descarga = TRUE, archivo = "theta_MAT_regiones.xlsx", ancho = 10, alto = 8)

# Análisis de la variable theta_MAT_E300 por región y edición
ARISTAS18.22.matematicas <- unionbases.longitudinal(ediciones = c(2018, 2022), bases = c("ARISTAS18.matematicas2", "ARISTAS22.matematicas"), y = "theta_MAT_300_50", x = c("regiones", "AlumnoGenero"))
longitudinal(datos = ARISTAS18.22.matematicas, ediciones = c(2018, 2022), y = "theta_MAT_300_50", x = "AlumnoGenero", digit = 1, plot = "density")

# Análisis de la variable theta_MAT_E300 por edición para los estudiantes de la región SUR
longitudinal(datos = ARISTAS18.22.matematicas |> dplyr::filter(regiones == "SUR"), ediciones = c(2018, 2022), y = "theta_MAT_300_50", x = NULL, digit = 3, plot = "density")

# Análisis de la variable theta_MAT_E300 por edición para los estudiantes de la región SUR y ESTE
longitudinal(datos = ARISTAS18.22.matematicas |> dplyr::filter(regiones \%in\% c("SUR", "ESTE")), ediciones = c(2018, 2022), y = "theta_MAT_300_50", x = "regiones", digit = 3, plot = "violin")

# -------------------------------------------------------------------------------------------------
# Comparación de nivel de desempeño de matemáticas
# -------------------------------------------------------------------------------------------------

# Unir las bases de las dos ediciones a una sola base de datos para anlizar Niveles_MAT por regiones
ARISTAS18.22.matematicas <- unionbases.longitudinal(ediciones = c(2018, 2022), bases = c("ARISTAS18.matematicas2", "ARISTAS22.matematicas"), y = "Niveles_MAT", x = "regiones")

# Análisis de la variable Niveles_MAT por edición a nivel nacional
longitudinal(datos = ARISTAS18.22.matematicas, ediciones = c(2018, 2022), y = "Niveles_MAT", digit = 2, plot = "bar")
longitudinal(datos = ARISTAS18.22.matematicas, ediciones = c(2018, 2022), y = "Niveles_MAT", digit = 3, plot = "bar", descarga = TRUE, archivo = "Niveles_MAT_ediciones.xlsx", ancho = 6, alto = 8)

# Análisis de la variable Niveles_MAT por regiones y edición
longitudinal(datos = ARISTAS18.22.matematicas, ediciones = c(2018, 2022), y = "Niveles_MAT", x = "regiones", digit = 3, plot = "bar")

# Análisis de la variable Niveles_MAT por edición para la región ESTE
longitudinal(datos = ARISTAS18.22.matematicas |> dplyr::filter(regiones == "ESTE"), ediciones = c(2018, 2022), y = "Niveles_MAT",  digit = 3, plot = "bar")

# Análisis de la variable Niveles_MAT por edición para la región ESTE y CENTRO
longitudinal(datos = ARISTAS18.22.matematicas |> dplyr::filter(regiones \%in\% c("ESTE", "CENTRO")), ediciones = c(2018, 2022), y = "Niveles_MAT", x = "regiones", digit = 3, plot = "bar")
}
