% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transversal.multiple.R
\name{transversal.multiple}
\alias{transversal.multiple}
\title{Esta función calcula las estimaciones de una variable de interés para una edición específica de ARISTAS para más de una variable de corte}
\usage{
transversal.multiple(
  data,
  y,
  x,
  IC = TRUE,
  digit = 3,
  plot = NULL,
  descarga = FALSE,
  archivo = NULL,
  ...
)
}
\arguments{
\item{data}{La base de datos que se quiere analizar, puede ser cualquiera de las bases de ARISTAS incluidas en el paquete o cualquiera otra base ARISTAS que no sea de uso público. Tenga en cuenta que las columnas que contengan los pesos replicados deben tener nombres: EST_W_REP_1, EST_W_REP_2, etc.}

\item{y}{El nombre de la variable de interés, ésta variable puede ser de tipo numérico o categórico.}

\item{x}{Vector de caracteres indicando el nombre de las variables de corte.}

\item{IC}{Indicar si se muestran las estimaciones e intervalos de confianza sobre la gráfica.}

\item{digit}{Especificar el número de dígitos decimales, por defecto son 3 dígitos decimales.}

\item{plot}{Las mismas opciones que las de la funci´on \code{transversal}.}

\item{descarga}{Indicar si desea descargar los resultados numéricos y la gráfica, debe especificar la ubicación. Al usar la opción \code{descarga = TRUE} se descargará un archivo excel con los resultados numéricos y pdf con las gráficas.}

\item{archivo}{Indicar el nombre del archivo excel que se descarga. Las gráficas se descargarán con el mismo nombre y extensión pdf en la misma ubicación que el archivo excel.}

\item{ancho}{El ancho en pulgadas de la gráfica, debe ser un vectro del mismo tamaño que el número de variables de corte.}

\item{alto}{El alto en pulgadas de la gráfica, debe ser un vectro del mismo tamaño que el número de variables de corte.}
}
\value{
\code{transversal.multiple} devuelve una lista de objetos correspondientes a estimacions y resultados de hipótesis de comparación:
\itemize{
\item \code{estimaciones}: La estimación, error estándar, intervalo de confianza de 95\% e información de los NA para la media de la variable de interés para cada valor de la variable de corte \code{x} y también a nivel global (para todos los datos ingresados en el argumento \code{data})
\item \code{contrastes_individuales}: Resultados de la prueba de hipótesis para la igualdad de medidas de \code{y} para cada combinación de 2 categorías de cada variable de corte.
\item \code{contraste_global}: Resultados de la prueba de hipótesis para la comparación múltiple de la media de \code{y} para todas las categorías de cada variable de corte.
}

Adicionalmente, también la función arrojará la gráfica que usuario haya elegido.
}
\description{
Esta función calcula las estimaciones de una variable de interés para una edición específica de ARISTAS para más de una variable de corte
}
\details{
Esta función arroja los mismos resultados de la función \code{transversal} permitiendo que el usuario ingrese más de una variable de corte.
}
\examples{

# Cambiar los valores de 0 en regiones por NA
ARISTAS18.matematicas$regiones[ARISTAS18.matematicas$regiones == 0] <- NA

# ------------------------------------------------------------------------------------------------------
# Análisis de theta_MAT_E300 con variables de corte regiones y AlumnoGenero15
# ------------------------------------------------------------------------------------------------------
transversal.multiple(data = ARISTAS18.matematicas, y = "theta_MAT_E300", x = c("regiones","AlumnoGenero15"))
transversal.multiple(data = ARISTAS18.matematicas, y = "theta_MAT_E300", x = c("regiones","AlumnoGenero15"), descarga = TRUE, plot = "density", archivo = "theta_MAT_E300.xlsx")
transversal.multiple(data = ARISTAS18.matematicas, y = "theta_MAT_E300", x = c("regiones","AlumnoGenero15"), descarga = TRUE, plot = "Lollipop", archivo = "theta_MAT_E300.xlsx", ancho = c(10,4), alto = c(8,5))

# ------------------------------------------------------------------------------------------------------
# Análisis de Niveles_MAT con variables de corte regiones y AlumnoGenero15
# ------------------------------------------------------------------------------------------------------
transversal.multiple(data = ARISTAS18.matematicas, y = "Niveles_MAT", x = c("regiones","AlumnoGenero15"))
transversal.multiple(data = ARISTAS18.matematicas, y = "Niveles_MAT", x = c("regiones","AlumnoGenero15"), descarga = TRUE, plot = "bar", archivo = "theta_MAT_E300.xlsx")
transversal.multiple(data = ARISTAS18.matematicas, y = "Niveles_MAT", x = c("regiones","AlumnoGenero15"), descarga = TRUE, plot = "bar", archivo = "theta_MAT_E300.xlsx", ancho = c(10,4), alto = c(6,9))
}
