% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transversal.R
\name{transversal}
\alias{transversal}
\title{Esta función calcula las estimaciones para una edición específica de ARISTAS}
\usage{
transversal(
  data,
  y,
  x = NULL,
  limits = NULL,
  digit = 3,
  IC = TRUE,
  plot = NULL,
  descarga = FALSE,
  archivo = NULL,
  ...
)
}
\arguments{
\item{data}{La base de datos que se quiere analizar, puede ser cualquiera de las bases de ARISTAS incluidas en el paquete o cualquiera otra base ARISTAS que no sea de uso público. Tenga en cuenta que las columnas que contengan los pesos replicados deben tener nombres: EST_W_REP_1, EST_W_REP_2, etc.}

\item{y}{El nombre de la variable de interés, ésta variable puede ser de tipo numérico o categórico.}

\item{x}{(Opcional) Nombre de la variable numérica o categórica que define los subgrupos para la comparación.}

\item{limits}{Vector de caracteres o vector numérico. En el caso de una variable numérica \code{x}, se debe proveer los valores límites para definir los subgrupos.
Por ejemplo, si \code{x} es la variable edad de los estudiantes, y se proveen los valores límites de 15 y 20, entonces se crearán tres subgrupos (intervalos): \eqn{(-\infty, 15]}, \eqn{(15, 20]} y \eqn{(20, \infty)}.
En el caso que \code{x} sea una variable categórica, los valores límites son opcionales.
Por ejemplo, si \code{x} es la variable región (con 5 valores para las 5 regiones de Uruguay), y se provee los valores 1 y 3, entonces la función creará tres subgrupos: \code{region==1}, \code{region==3} y la unión de las demás regiones.
Si no se provee valores límites para una variable \code{x} categórica, entonces la función comparará la variable de interés \code{y} en todas las categorías de \code{x}.}

\item{digit}{Especificar el número de dígitos decimales, por defecto son 3 dígitos decimales.}

\item{IC}{Indicar si se muestran las estimaciones e intervalos de confianza sobre la gráfica.}

\item{plot}{Puede elegir entre \code{histogram}, \code{density}, \code{violin}, \code{Lollipop}  para una variable numérica, la opción \code{violin} incluirá también la gráfica de cajas. Para una variable categórica, puede usar la opción \code{bar}.}

\item{descarga}{Indicar si desea descargar los resultados numéricos y la gráfica, debe especificar la ubicación. Al usar la opción \code{descarga = TRUE} se descargará un archivo excel con los resultados numéricos y un pdf con la gráfica.}

\item{archivo}{Indicar el nombre del archivo excel que se descarga. La gráfica, si hay, se descargará con el mismo nombre y extensión pdf en la misma ubicación que el archivo excel.}

\item{ancho}{El ancho en pulgadas de la gráfica.}

\item{alto}{El alto en pulgadas de la gráfica.}
}
\value{
\code{transversal} devuelve una lista de objetos correspondientes a estimacions y resultados de hipótesis de comparación:
\itemize{
\item \code{estimaciones}: La estimación, error estándar, intervalo de confianza de 95\% e información de los NA para la media de la variable de interés para cada valor de la variable de corte \code{x} y también a nivel global (para todos los datos ingresados en el argumento \code{data})
\item \code{contraste.individual}: Resultados de la prueba de hipótesis para la igualdad de medidas para cada combinación de 2 categorías de la variable de corte (de las categorías especificados por el usuario en el argumento \code{limits}). Estos resultas solo se muestran cuando el usuario no especifica la variable de corte.
\item \code{contraste.global}: Resultados de la prueba de hipótesis para la comparación múltiple de la media de \code{y} para todas las categorías especificados por el usuario en el argumento \code{limits}. Estos resultas solo se muestran cuando el usuario no especifica la variable de corte.
}

Adicionalmente, también la función arrojará la gráfica que usuario haya elegido.
}
\description{
Esta función calcula las estimaciones para una edición específica de ARISTAS
}
\details{
Cuando la variable de interés \code{y} es una variable numérica (puntaje en pruebas), y el usuario ingresa la varible de corte \code{x} con o sin valores en el argumento \code{limits},
la función realizará la prueba \eqn{t} para cada pareja de categorías de \code{x} con la función \code{survey::svyttest} con base en un diseño muestral de pesos réplica definido con \code{survey::svrepdesign},
y estos resultados se encuentran en el resultado \code{contraste.individual}. La salida de la función incluye la estimación, error estándar, intervalo de confianza de 95\% para la media de la variable \code{y} para cada categoría de \code{x} y a nivel global
También calcula el número de datos, número y porcentaje de NA para cada categoría de \code{x} y a nivel global
De la misma forma, la función llevará a cabo la prueba global \eqn{F} de igualdad de medias  para todas las categorías de la variable \code{x}.

Cuando la variable de interés \code{y} es una variable numérica (puntaje en pruebas), y el usuario no ingresa la varible de corte \code{x},
la función solo calculará la estimación, error estándar, intervalo de confianza de 95\%, cantidad de datos, cantidad y porcentaje de NA para la media de la variable \code{y} para todos los datos ingresados en \code{data}.

Cuando la variable de interés \code{y} es una variable categórica (nivel de desempeño), y el usuario ingresa la varible de corte \code{x} con o sin valores en el argumento \code{limits},
la función realizará la prueba chi-cuadrado para cada pareja de categorías de \code{x} con la función \code{survey::svytable} para la independencia entre los diferentes niveles de desempeño y las dos categorías de la variable \code{x},
lo anterior con base en un diseño muestral de pesos réplica definido con \code{survey::svrepdesign},
y estos resultados se encuentran en \code{contraste.individual}.
También calcula el número de datos, número y porcentaje de NA para cada categoría de \code{x} y a nivel global
De la misma forma, la función llevará a cabo la prueba global de independencia de chi-cuadrado entre los diferentes niveles de desempeño y todas las categorías de la variable \code{x}.
Cuando el usuario no especifica la variable de corte \code{x}, el resultado \code{contraste.individual} no está en la salida de la función.

Cuando la variable de interés \code{y} es una variable categórica (nivel de desempeño), y el usuario no ingresa la varible de corte \code{x},
la función solo calculará la estimación, error estándar e intervalo de confianza de 95\% para la variable \code{y} para todos los datos ingresados en \code{data}.

La opción de \code{plot} para una variable de interés \code{y} numérica admite \code{histogram}  (el cual hace uso de la función \code{survey::svyhist}), \code{density}, \code{violin} y \code{Lollipop}, mientras que
para una variable de interés \code{y} categórica admite la opción de \code{bar}.

Las estimaciones se obtuvieron utilizando la función \code{svymean} del paquete \code{survey} apoyado en un diseño muestral de pesos réplica creado con \code{svrepdesign}
}
\examples{
# Cambiar los valores de 0 en regiones por NA
ARISTAS18.matematicas$regiones[ARISTAS18.matematicas$regiones == 0] <- NA

# ------------------------------------------------------------------------------------------------------
# Variable de interés numérica y con variable de corte corte x categórica (regiones y AlumnoGenero15)
# ------------------------------------------------------------------------------------------------------

# Análisis de la variable theta_MAT_E300 con la gráfica de histograma
transversal(data = ARISTAS18.matematicas, y = "theta_MAT_E300", plot = "histogram")
transversal(data = ARISTAS18.matematicas, y = "theta_MAT_E300", x = "AlumnoGenero15", plot = "histogram")
transversal(data = ARISTAS18.matematicas, y = "theta_MAT_E300", x = "regiones", digit = 3, plot = "histogram")
transversal(data = ARISTAS18.matematicas, y = "theta_MAT_E300", x = "regiones", digit = 3, plot = NULL, descarga = TRUE, archivo = "theta_MAT_regiones.xlsx")
transversal(data = ARISTAS18.matematicas, y = "theta_MAT_E300", x = "regiones", digit = 4, plot = "histogram", descarga = TRUE, archivo = "theta_MAT_regiones.xlsx", ancho = 6, alto = 8)

# Análisis de la variable theta_MAT_E300 con la gráfica de densidad
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", x = "regiones", plot = "density")
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", x = "regiones", IC = FALSE, plot = "density")
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", x = "AlumnoGenero15", plot = "density")
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", x = "AlumnoGenero15", IC = FALSE, plot = "density")
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", IC = TRUE, plot = NULL, descarga = TRUE, archivo = "theta_MAT_E300.xlsx")
# Al no especificar la dimensión de la gráfica, el pdf guardará la gráfica con las mismas dimensiones que aparece en el panel de Plots
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", IC = FALSE, plot = "density", descarga = TRUE, archivo = "theta_MAT_E300.xlsx")

# Análisis de la variable theta_MAT_E300 con la gráfica de violin
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", x = "regiones", plot = "violin")
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", x = "regiones", IC = FALSE, plot = "violin")
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", x = "AlumnoGenero15", IC = TRUE, plot = "violin")
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", x = "AlumnoGenero15", IC = FALSE, plot = "violin")
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", IC = TRUE, plot = "violin")
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", IC = FALSE, plot = "violin", descarga = TRUE, archivo = "theta_MAT_E300.xlsx")

# Análisis de la variable theta_MAT_E300 con la gráfica de Lollipop
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", x = "regiones", IC = TRUE, plot = "Lollipop")
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", x = "regiones", IC = FALSE, plot = "Lollipop")
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", x = "AlumnoGenero15", IC = TRUE, plot = "Lollipop", descarga = TRUE, archivo = "theta_MAT_E300.xlsx", ancho = 6, alto = 8)
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", x = "AlumnoGenero15", IC = FALSE, plot = "Lollipop")
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", IC = TRUE, plot = "Lollipop")
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", IC = FALSE, plot = "Lollipop")

# ---------------------------------------------------------------------------------------
# Variable de interés numérica y con variable de corte corte corte x numérica (EDAD)
# ---------------------------------------------------------------------------------------
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", x = "EDAD", limits = c(15), IC = TRUE, plot = "histogram")
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", x = "EDAD", limits = c(15, 20, 22), IC = TRUE, plot = "density")
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", x = "EDAD", limits = c(15, 20, 22), IC = TRUE, plot = "violin")
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", x = "EDAD", limits = c(15, 20, 22), IC = TRUE, plot = "Lollipop")

# --------------------------------------------------------------------------------------------------------------
# Variable de interés numérica y con varaible de corte x categórica (regiones) con categorías especificadas
# --------------------------------------------------------------------------------------------------------------
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", x = "regiones", limits = c("2"), IC = TRUE, plot = "histogram")
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", x = "regiones", limits = c("2"), IC = TRUE, plot = "density")
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", x = "regiones", limits = c("2", "4"), IC = TRUE, plot = "violin")
transversal(data=ARISTAS18.matematicas, y = "theta_MAT_E300", x = "regiones", limits = c("2", "4"), IC = TRUE, plot = "Lollipop")

# ----------------------------------------------------------------------
# Variable de interés y categórica con variable de corte x categórica
# ----------------------------------------------------------------------

# Gráfica de barras
transversal(data=ARISTAS18.matematicas, y = "Niveles_MAT", x = "AlumnoGenero15", plot = "bar", descarga = TRUE, archivo = "Niveles_MAT.xlsx")
transversal(data=ARISTAS18.matematicas, y = "Niveles_MAT", x = "regiones", plot  = "bar")
transversal(data=ARISTAS18.matematicas, y = "Niveles_MAT", x = "regiones", limits =  c("1", "5"), plot  = "bar")
transversal(data=ARISTAS18.matematicas, y = "Niveles_MAT", plot  = "bar")
}
