% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unionbases.longitudinal.R
\name{unionbases.longitudinal}
\alias{unionbases.longitudinal}
\title{Esta función une bases de ARISTAS de diferentes ediciones para poder ser analizado con la fucnión \code{longitudinal}.}
\usage{
unionbases.longitudinal(ediciones, bases, y, x = NULL)
}
\arguments{
\item{ediciones}{Vector que indica para cuáles ediciones se quiere unir las bases de datos.}

\item{y}{Nombre de la variable de interés \code{y}.}

\item{x}{(Opcional) Nombre de la variable de corte \code{x}.}

\item{datos}{Vector con el nombre de las bases que se quiere unir, con el mismo orden del parámetro \code{ediciones}.}
}
\value{
\code{unionbases.longitudinal}: devuelve una base de datos uniendo las bases de datos indicados por el usuario, con una variable de nombre \code{edicion} indicando el año.
}
\description{
Esta función une bases de ARISTAS de diferentes ediciones para poder ser analizado con la fucnión \code{longitudinal}.
}
\details{
Las variables (variable de interés 'y', la variable de corte 'x') de las bases de las diferentes ediciones deben:
(1) tener el mismo nombre y (2) tener los mismos valores.
Los pesos de estudiantes deben llamarse 'peso_MEst',  los pesos réplicas deben llamarse 'EST_W_REP_1’, 'EST_W_REP_2’, etc.
}
\examples{

# Convertir las variables numéricas a categóricas
ARISTAS18.matematicas$regiones <- as.character(ARISTAS18.matematicas$regiones)
# Unificar las categorías de ARISTAS 2018 según las categorías de 2022
ARISTAS18.matematicas2 <- ARISTAS18.matematicas |>
  dplyr::mutate(regiones = dplyr::case_when(
    regiones == 1 ~ "SUR",
    regiones == 2 ~ "OESTE",
    regiones == 3 ~ "NORTE",
    regiones == 4 ~ "ESTE",
    regiones == 5 ~ "CENTRO",
    TRUE ~ NA_character_  # Para asegurarse de que otros valores no definidos se traten como NA
    )
  )

# Unificar el nombre de la variable theta_MAT_300_50
ARISTAS18.matematicas2 <- ARISTAS18.matematicas2 |>
  dplyr::rename(theta_MAT_300_50 = theta_MAT_E300,
                AlumnoGenero = AlumnoGenero15)

ARISTAS18.22.matematicas <- unionbases.longitudinal(ediciones = c(2018, 2022), bases = c("ARISTAS18.matematicas2", "ARISTAS22.matematicas"), y = "theta_MAT_300_50", x = "regiones")
ARISTAS18.22.matematicas <- unionbases.longitudinal(ediciones = c(2018, 2022), bases = c("ARISTAS18.matematicas2", "ARISTAS22.matematicas"), y = "theta_MAT_300_50", x = c("regiones", "AlumnoGenero"))

}
